커뮤니티 게시판 삭제 구현(글 및 댓글 수정기능 구현필요)

+++ backend/src/main/java/com/bitc/finalproject/controller/BoardControl.java
@@ -40,8 +40,11 @@ public class BoardControl {
     }
 
     @RequestMapping(value = "/board/{boardPk}", method = RequestMethod.DELETE)
-    public Object boardDelete(@PathVariable("boardPk") int boardPk) throws Exception {
-        boardService.deleteBoard(boardPk);
+    public Object boardDelete(@PathVariable("boardPk") int boardPk, @RequestParam("boardWriterId") String boardWriterId) throws Exception {
+
+        if (boardWriterId.equals("testUserId")) {
+            boardService.deleteBoard(boardPk);
+        }
 
         return "redirect:/board";
     }
